bugfix: uint used in sparse_sd_vector and succinct_bitvector

+++ internal/sparse_sd_vector.hpp
@@ -43,7 +43,7 @@ public:
 
 		bit_vector bv(b.size());
 
-		for(uint i=0;i<b.size();++i)
+		for(uint64_t i=0;i<b.size();++i)
 			bv[i] = b[i];
 
 		sdv = sd_vector<>(bv);
